Refactoring: Introduce InterfaceSkeleton component and use it in the integration package
