+++ packages/dashboard/src/app/index.js
@@ -14,219 +14,5 @@
  * limitations under the License.
  */
 
-/**
- * External dependencies
- */
-import {
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  usePrevious,
-} from '@web-stories-wp/react';
-import { StyleSheetManager, ThemeProvider } from 'styled-components';
-import stylisRTLPlugin from 'stylis-plugin-rtl';
-import PropTypes from 'prop-types';
-import { __, sprintf } from '@web-stories-wp/i18n';
-import { trackScreenView } from '@web-stories-wp/tracking';
-import {
-  theme as externalDesignSystemTheme,
-  lightMode,
-  ThemeGlobals,
-  useSnackbar,
-  SnackbarProvider,
-  Snackbar,
-  ModalGlobalStyle,
-} from '@web-stories-wp/design-system';
-
-/**
- * Internal dependencies
- */
-import { GlobalStyle } from '../theme';
-import { KeyboardOnlyOutline } from '../utils';
-import {
-  APP_ROUTES,
-  NESTED_APP_ROUTES,
-  ROUTE_TITLES,
-  ADMIN_TITLE,
-} from '../constants';
-
-import { AppFrame, LeftRail, NavProvider, PageContent } from '../components';
-import ApiProvider from './api/apiProvider';
-import { ConfigProvider } from './config';
-import { Route, RouterProvider, matchPath, useRouteHistory } from './router';
-import {
-  EditorSettingsView,
-  ExploreTemplatesView,
-  MyStoriesView,
-  TemplateDetailsView,
-} from './views';
-import useApi from './api/useApi';
-import useApiAlerts from './api/useApiAlerts';
-
-const AppContent = () => {
-  const {
-    state: {
-      currentPath,
-      queryParams: { id: templateId },
-    },
-    actions: { push },
-  } = useRouteHistory();
-  const { currentTemplate, addInitialFetchListener } = useApi(
-    ({
-      actions: {
-        storyApi: { addInitialFetchListener },
-      },
-      state: {
-        templates: { templates },
-      },
-    }) => ({
-      currentTemplate:
-        templateId !== undefined ? templates[templateId]?.title : undefined,
-      addInitialFetchListener,
-    })
-  );
-  const isFirstLoadOnMyStories = useRef(currentPath === APP_ROUTES.DASHBOARD);
-  const [isRedirectComplete, setIsRedirectComplete] = useState(
-    !isFirstLoadOnMyStories.current
-  );
-
-  // Direct user to templates on first load if they
-  // have no stories created.
-  useEffect(() => {
-    return addInitialFetchListener?.((storyStatuses) => {
-      if (storyStatuses?.all <= 0 && isFirstLoadOnMyStories.current) {
-        push(APP_ROUTES.TEMPLATES_GALLERY);
-      }
-      setIsRedirectComplete(true);
-    });
-  }, [addInitialFetchListener, push, currentPath]);
-
-  const fullPath = useMemo(() => {
-    return currentPath.includes(APP_ROUTES.TEMPLATE_DETAIL) &&
-      templateId &&
-      currentTemplate
-      ? `${currentPath}/${templateId}`
-      : currentPath;
-    // Disable reason: avoid sending duplicate tracking events.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentPath, currentTemplate]);
-
-  useEffect(() => {
-    if (!isRedirectComplete) {
-      return;
-    }
-
-    if (currentPath.includes(APP_ROUTES.TEMPLATE_DETAIL) && !currentTemplate) {
-      return;
-    }
-
-    let dynamicPageTitle = ROUTE_TITLES[currentPath] || ROUTE_TITLES.DEFAULT;
-    if (currentPath.includes(APP_ROUTES.TEMPLATE_DETAIL)) {
-      dynamicPageTitle = sprintf(
-        /* translators: %s: Template name. */
-        __('Template: %s', 'web-stories'),
-        currentTemplate
-      );
-    }
-
-    document.title = sprintf(
-      /* translators: Admin screen title. 1: Admin screen name, 2: Network or site name. */
-      __('%1$s \u2039 %2$s \u2212 WordPress', 'web-stories'),
-      dynamicPageTitle,
-      ADMIN_TITLE
-    );
-
-    trackScreenView(dynamicPageTitle);
-
-    // Disable reason: avoid sending duplicate tracking events.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [fullPath, isRedirectComplete]);
-
-  const hideLeftRail = matchPath(
-    currentPath,
-    NESTED_APP_ROUTES.TEMPLATES_GALLERY_DETAIL
-  );
-
-  useApiAlerts();
-  const { clearSnackbar, removeSnack, placement, currentSnacks } =
-    useSnackbar();
-
-  // if the current path changes clear the snackbar
-  const prevPath = usePrevious(currentPath);
-
-  useEffect(() => {
-    if (currentPath !== prevPath) {
-      clearSnackbar();
-    }
-  }, [clearSnackbar, currentPath, prevPath]);
-
-  return (
-    <>
-      <AppFrame>
-        {!hideLeftRail && <LeftRail />}
-        <PageContent fullWidth={hideLeftRail}>
-          <Route
-            exact
-            path={APP_ROUTES.DASHBOARD}
-            component={<MyStoriesView />}
-          />
-          <Route
-            exact
-            path={APP_ROUTES.TEMPLATES_GALLERY}
-            component={<ExploreTemplatesView />}
-          />
-          <Route
-            path={NESTED_APP_ROUTES.TEMPLATES_GALLERY_DETAIL}
-            component={<TemplateDetailsView />}
-          />
-          <Route
-            path={APP_ROUTES.EDITOR_SETTINGS}
-            component={<EditorSettingsView />}
-          />
-        </PageContent>
-      </AppFrame>
-      <Snackbar.Container
-        notifications={currentSnacks}
-        onRemove={removeSnack}
-        placement={placement}
-        max={1}
-      />
-    </>
-  );
-};
-
-function App({ config }) {
-  const { isRTL } = config;
-  const activeTheme = {
-    ...externalDesignSystemTheme,
-    colors: lightMode,
-  };
-  return (
-    <StyleSheetManager stylisPlugins={isRTL ? [stylisRTLPlugin] : []}>
-      <ThemeProvider theme={activeTheme}>
-        <ThemeGlobals.Styles />
-        <ModalGlobalStyle />
-        <ConfigProvider config={config}>
-          <ApiProvider>
-            <NavProvider>
-              <RouterProvider>
-                <SnackbarProvider>
-                  <GlobalStyle />
-                  <KeyboardOnlyOutline />
-                  <AppContent />
-                </SnackbarProvider>
-              </RouterProvider>
-            </NavProvider>
-          </ApiProvider>
-        </ConfigProvider>
-      </ThemeProvider>
-    </StyleSheetManager>
-  );
-}
-
-App.propTypes = {
-  config: PropTypes.object.isRequired,
-};
-
-export default App;
+export * from './api';
+export * from './config';
