+++ packages/wp-dashboard/src/index.js
@@ -25,7 +25,7 @@ import './publicPath';
 /**
  * External dependencies
  */
-import Dashboard from '@web-stories-wp/dashboard';
+import Dashboard, { InterfaceSkeleton } from '@web-stories-wp/dashboard';
 import { setAppElement } from '@web-stories-wp/design-system';
 import { StrictMode, render } from '@web-stories-wp/react';
 import { FlagsProvider } from 'flagged';
@@ -58,7 +58,9 @@ const initialize = async (id, config, flags) => {
   render(
     <FlagsProvider features={flags}>
       <StrictMode>
-        <Dashboard config={config} />
+        <Dashboard config={config}>
+          <InterfaceSkeleton />
+        </Dashboard>
       </StrictMode>
     </FlagsProvider>,
     appElement
